work on parser rework
